@@ -12,26 +12,26 @@
         }
     ],
     "require": {
-        "php": ">=5.5.9",
-        "symfony/framework-bundle": "~2.7|~3.0|~4.0|^5",
-        "symfony/console": "~2.7|~3.0|~4.0|^5",
-        "symfony/dependency-injection": "~2.7|~3.0|~4.0|^5",
-        "doctrine/doctrine-bundle": ">=1.6",
+        "php": ">=8.0",
+        "symfony/framework-bundle": "~4.0|~5.0|~6.0",
+        "symfony/console": "~4.0|~5.0|~6.0",
+        "symfony/dependency-injection": "~4.0|~5.0|~6.0",
+        "doctrine/doctrine-bundle": ">=2.0",
         "guzzlehttp/guzzle": "~6.0",
 		"guzzlehttp/promises": ">=1.0"
     },
     "require-dev": {
-        "satooshi/php-coveralls": "^1.0",
+        "php-coveralls/php-coveralls": "^2.5.2",
         "phpunit/phpunit": "*",
         "doctrine/orm": "^2.5",
-        "symfony/yaml": "~2.7|~3.0|^4|^5"
+        "symfony/yaml": "~4.0|~5.0|~6.0"
     },
     "autoload": {
         "psr-4": { "Bordeux\\Bundle\\GeoNameBundle\\": "" }
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.0.x-dev"
+            "dev-master": "3.0.x-dev"
         }
     }
 }
