@@ -3,6 +3,8 @@
 namespace Bordeux\Bundle\GeoNameBundle\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
+use phpDocumentor\Reflection\Types\Integer;
+use Stringable;
 
 /**
  * GeoName
@@ -10,9 +12,9 @@ use Doctrine\ORM\Mapping as ORM;
  * @ORM\Table(name="geo__country" ,indexes={
  *     @ORM\Index(name="geoname_country_search_idx", columns={"name", "iso"})
  * })
- * @ORM\Entity(repositoryClass="Bordeux\Bundle\GeoNameBundle\Repository\CountryRepository")
+ * @ORM\Entity()
  */
-class Country
+class Country implements Stringable
 {
     /**
      * @var int
@@ -21,28 +23,28 @@ class Country
      * @ORM\Id
      * @ORM\GeneratedValue(strategy="NONE")
      */
-    protected $id;
+    protected int $id;
 
     /**
      * @var string
      *
      * @ORM\Column(name="iso", type="string", length=2, nullable=false)
      */
-    protected $iso;
+    protected string $iso;
 
     /**
      * @var string
      *
      * @ORM\Column(name="iso3", type="string", length=3, nullable=false)
      */
-    protected $iso3;
+    protected string $iso3;
 
     /**
      * @var integer
      *
      * @ORM\Column(name="iso_numeric", type="integer", length=3, nullable=false)
      */
-    protected $isoNumeric;
+    protected int $isoNumeric;
 
 
     /**
@@ -50,7 +52,7 @@ class Country
      *
      * @ORM\Column(name="fips", type="string", length=2, nullable=true)
      */
-    protected $fips;
+    protected ?string $fips;
 
 
     /**
@@ -58,14 +60,14 @@ class Country
      *
      * @ORM\Column(name="name", type="string", length=255, nullable=false)
      */
-    protected $name;
+    protected string $name;
 
     /**
      * @var string
      *
      * @ORM\Column(name="capital", type="string", length=255, nullable=true)
      */
-    protected $capital;
+    protected ?string $capital;
 
 
     /**
@@ -73,45 +75,43 @@ class Country
      *
      * @ORM\Column(name="area", type="bigint", nullable=false)
      */
-    protected $area;
+    protected int $area;
 
     /**
-     * @var integer
+     * @var int
      *
      * @ORM\Column(name="population", type="bigint", nullable=false)
      */
-    protected $population;
-
+    protected int $population;
 
     /**
      * @var string
      *
      * @ORM\Column(name="tld", type="string", length=15, nullable=true)
      */
-    protected $tld;
+    protected ?string $tld;
 
     /**
      * @var string
      *
      * @ORM\Column(name="currency", type="string", length=3, nullable=true)
      */
-    protected $currency;
-
+    protected ?string $currency;
 
     /**
      * @var string
      *
      * @ORM\Column(name="currency_name", type="string", length=50, nullable=true)
      */
-    protected $currencyName;
+    protected ?string $currencyName;
 
 
     /**
-     * @var integer
+     * @var int
      *
      * @ORM\Column(name="phone_prefix", type="integer", nullable=true)
      */
-    protected $phonePrefix;
+    protected ?int $phonePrefix;
 
 
     /**
@@ -119,23 +119,22 @@ class Country
      *
      * @ORM\Column(name="postal_format", type="text", nullable=true)
      */
-    protected $postalFormat;
-
+    protected ?string $postalFormat;
 
     /**
      * @var string
      *
      * @ORM\Column(name="postal_regex", type="text", nullable=true)
      */
-    protected $postalRegex;
+    protected ?string $postalRegex;
 
 
     /**
      * @var array
      *
-     * @ORM\Column(name="languages", type="json_array", nullable=true)
+     * @ORM\Column(name="languages", type="json", nullable=true)
      */
-    protected $languages;
+    protected ?array $languages;
 
     /**
      * @var GeoName
@@ -143,365 +142,312 @@ class Country
      * @ORM\ManyToOne(targetEntity="Bordeux\Bundle\GeoNameBundle\Entity\GeoName")
      * @ORM\JoinColumn(name="geoname_id", referencedColumnName="id", nullable=true)
      */
-    protected $geoName;
+    protected ?GeoName $geoName;
 
     /**
-     * Country constructor.
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @param int $id
-     */
-    public function __construct($id = null)
-    {
-        $this->id = $id;
-    }
-
-    /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
      * @return int
      */
-    public function getId()
+    public function getId(): int
     {
         return $this->id;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @param int $id
-     * @return Country
-     */
-    public function setId($id)
-    {
-        $this->id = $id;
-        return $this;
-    }
-
-    /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
      * @return string
      */
-    public function getIso()
+    public function getIso(): string
     {
         return $this->iso;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
      * @param string $iso
-     * @return Country
+     * @return $this
      */
-    public function setIso($iso)
+    public function setIso(string $iso): self
     {
         $this->iso = $iso;
         return $this;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
      * @return string
      */
-    public function getIso3()
+    public function getIso3(): string
     {
         return $this->iso3;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
      * @param string $iso3
-     * @return Country
+     * @return $this
      */
-    public function setIso3($iso3)
+    public function setIso3(string $iso3): self
     {
         $this->iso3 = $iso3;
         return $this;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
      * @return int
      */
-    public function getIsoNumeric()
+    public function getIsoNumeric(): int
     {
         return $this->isoNumeric;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
      * @param int $isoNumeric
-     * @return Country
+     * @return $this
      */
-    public function setIsoNumeric($isoNumeric)
+    public function setIsoNumeric(int $isoNumeric): self
     {
         $this->isoNumeric = $isoNumeric;
         return $this;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @return string
+     * @return string|null
      */
-    public function getFips()
+    public function getFips(): ?string
     {
         return $this->fips;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @param string $fips
-     * @return Country
+     * @param string|null $fips
+     * @return $this
      */
-    public function setFips($fips)
+    public function setFips(?string $fips): self
     {
         $this->fips = $fips;
         return $this;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
      * @return string
      */
-    public function getName()
+    public function getName(): string
     {
         return $this->name;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
      * @param string $name
-     * @return Country
+     * @return $this
      */
-    public function setName($name)
+    public function setName(string $name): self
     {
         $this->name = $name;
         return $this;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
      * @return int
      */
-    public function getArea()
+    public function getArea(): int
     {
         return $this->area;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @return string
+     * @param int $area
+     * @return $this
      */
-    public function getCapital()
+    public function setArea(int $area): self
+    {
+        $this->area = (int)$area;
+        return $this;
+    }
+
+
+    /**
+     * @return string|null
+     */
+    public function getCapital(): ?string
     {
         return $this->capital;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @param string $capital
-     * @return Country
+     * @param string|null $capital
+     * @return $this
      */
-    public function setCapital($capital)
+    public function setCapital(?string $capital): self
     {
         $this->capital = $capital;
         return $this;
     }
 
 
-
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @param int $area
-     * @return Country
-     */
-    public function setArea($area)
-    {
-        $this->area =  (int) $area;
-        return $this;
-    }
-
-    /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
      * @return int
      */
-    public function getPopulation()
+    public function getPopulation(): int
     {
         return $this->population;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
      * @param int $population
-     * @return Country
+     * @return $this
      */
-    public function setPopulation($population)
+    public function setPopulation(int $population): self
     {
-        $this->population = (int)  $population;
+        $this->population = (int)$population;
         return $this;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @return string
+     * @return string|null
      */
-    public function getTld()
+    public function getTld(): ?string
     {
         return $this->tld;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @param string $tld
-     * @return Country
+     * @param string|null $tld
+     * @return $this
      */
-    public function setTld($tld)
+    public function setTld(?string $tld): self
     {
         $this->tld = $tld;
         return $this;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @return string
+     * @return string|null
      */
-    public function getCurrency()
+    public function getCurrency(): ?string
     {
         return $this->currency;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @param string $currency
-     * @return Country
+     * @param string|null $currency
+     * @return $this
      */
-    public function setCurrency($currency)
+    public function setCurrency(?string $currency): self
     {
         $this->currency = $currency;
         return $this;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @return string
+     * @return string|null
      */
-    public function getCurrencyName()
+    public function getCurrencyName(): ?string
     {
         return $this->currencyName;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
      * @param string $currencyName
-     * @return Country
+     * @return $this
      */
-    public function setCurrencyName($currencyName)
+    public function setCurrencyName(string $currencyName): self
     {
         $this->currencyName = $currencyName;
         return $this;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @return int
+     * @return int|null
      */
-    public function getPhonePrefix()
+    public function getPhonePrefix(): ?int
     {
         return $this->phonePrefix;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @param int $phonePrefix
-     * @return Country
+     * @param int|null $phonePrefix
+     * @return $this
      */
-    public function setPhonePrefix($phonePrefix)
+    public function setPhonePrefix(?int $phonePrefix): self
     {
         $this->phonePrefix = $phonePrefix;
         return $this;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @return string
+     * @return string|null
      */
-    public function getPostalFormat()
+    public function getPostalFormat(): ?string
     {
         return $this->postalFormat;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @param string $postalFormat
-     * @return Country
+     * @param $postalFormat
+     * @return $this
      */
-    public function setPostalFormat($postalFormat)
+    public function setPostalFormat(?string $postalFormat): self
     {
         $this->postalFormat = $postalFormat;
         return $this;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @return string
+     * @return string|null
      */
-    public function getPostalRegex()
+    public function getPostalRegex(): ?string
     {
         return $this->postalRegex;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @param string $postalRegex
-     * @return Country
+     * @param string|null $postalRegex
+     * @return $this
      */
-    public function setPostalRegex($postalRegex)
+    public function setPostalRegex(?string $postalRegex): self
     {
         $this->postalRegex = $postalRegex;
         return $this;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @return array
+     * @return array|null
      */
-    public function getLanguages()
+    public function getLanguages(): ?array
     {
         return $this->languages;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @param array $languages
-     * @return Country
+     * @param array|null $languages
+     * @return $this
      */
-    public function setLanguages($languages)
+    public function setLanguages(?array $languages): self
     {
         $this->languages = $languages;
         return $this;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @return GeoName
+     * @return GeoName|null
      */
-    public function getGeoName()
+    public function getGeoName(): ?GeoName
     {
         return $this->geoName;
     }
 
     /**
-     * @author Chris Bednarczyk <chris@tourradar.com>
-     * @param GeoName $geoName
-     * @return Country
+     * @param GeoName|null $geoName
+     * @return $this
      */
-    public function setGeoName($geoName)
+    public function setGeoName(?GeoName $geoName): self
     {
         $this->geoName = $geoName;
         return $this;
     }
 
-
-
-
-
-
-
+    /**
+     * @return string
+     */
+    public function __toString(): string
+    {
+        return $this->getName();
+    }
 }
 
